src/main/java/git/w1shm4st3r/recipeapp/converter/CategoryCommandToCategory.java
package git.w1shm4st3r.recipeapp.converter;

import git.w1shm4st3r.recipeapp.commands.CategoryCommand;
import git.w1shm4st3r.recipeapp.domain.Category;
import lombok.Synchronized;
import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;

@Component
public class CategoryCommandToCategory implements Converter<CategoryCommand, Category> {

    @Synchronized
    @Nullable
    @Override
    public Category convert(CategoryCommand source) {
        if (source == null) {
            return null;
        }
        final Category category = new Category();
        category.setId(source.getId());
        category.setDescription(source.getDescription());
        return category;
    }
}
